Fix building warnings

+++ src/xrGame/ai/monsters/basemonster/base_monster_debug.cpp
@@ -887,7 +887,7 @@ void   CBaseMonster::add_debug_info (debug::text_tree& root_s)
 	for ( SoundIterator i=sound().objects().begin(), e=sound().objects().end();
 		  i!=e; ++i )
 	{
-		object_count += (*i).second.second->m_sounds.size();
+		object_count += (u32)(*i).second.second->m_sounds.size();
 	}
 
 	TextTree& now_playing_s = sound_player_s.add_line("Objects", object_count);
